[cpp] Disambiguate std::abs on tv_usec in 10_bicycle_03
On Linux glibc tv_usec is long, so std::abs(now.tv_usec - ...) has no
exact overload and resolves ambiguously under llvm-22 (on macOS tv_usec
is int, so it compiled). Cast the difference to double so it matches
double abs(double) from <cmath> unambiguously on every platform.

+++ regression/esbmc-cpp/unix/10_bicycle_03/EmbeddedPC.cpp
@@ -140,7 +140,7 @@ void EmbeddedPC::wheelTurn()
 	struct timeval now;
 	gettimeofday(&now, NULL);
 
-	float time = std::abs(now.tv_usec - _lastWheelTurn.tv_usec)/1000000;
+	float time = std::abs((double)(now.tv_usec - _lastWheelTurn.tv_usec))/1000000;
 	__ESBMC_assume(time!=0);
 	// each wheel turn, 1m is moved
 	_dist += 1;
